docker-requirements.txt
numpy==1.26.4
pandas==2.2.2
Django==5.0.6
djangorestframework==3.15.1
